requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
psycopg2-binary>=2.9.9,<3.0